# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.City:
    properties:
      coordinates:
        $ref: '#/definitions/domain.Coordinates'
      created_at:
        type: string
      districts:
        items:
          $ref: '#/definitions/domain.District'
        type: array
      id:
        type: integer
      name:
        type: string
      region:
        $ref: '#/definitions/domain.Region'
      region_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  domain.Coordinates:
    properties:
      city_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      updated_at:
        type: string
    type: object
...
  domain.District:
    properties:
      city:
        $ref: '#/definitions/domain.City'
      city_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      microdistricts:
        items:
          $ref: '#/definitions/domain.Microdistrict'
        type: array
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.Microdistrict:
    properties:
      created_at:
        type: string
      district:
        $ref: '#/definitions/domain.District'
      district_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.Region:
    properties:
      cities:
        items:
          $ref: '#/definitions/domain.City'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для платформы аренды квартир Renti.kz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Renti.kz API
  version: "1.0"
paths:
...
  /locations/cities:
    get:
      consumes:
      - application/json
      description: Получает список всех городов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.City'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение всех городов
      tags:
      - locations
  /locations/cities/{id}:
    get:
      consumes:
      - application/json
      description: Получает информацию о городе по его идентификатору
      parameters:
      - description: ID города
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.City'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение города по ID
      tags:
      - locations
  /locations/cities/{id}/districts:
    get:
      consumes:
      - application/json
      description: Получает список районов для указанного города
      parameters:
      - description: ID города
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.District'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение районов города
      tags:
      - locations
  /locations/districts:
    get:
      consumes:
      - application/json
      description: Получает список всех районов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.District'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение всех районов
      tags:
      - locations
  /locations/districts/{id}:
    get:
      consumes:
      - application/json
      description: Получает информацию о районе по его идентификатору
      parameters:
      - description: ID района
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.District'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение района по ID
      tags:
      - locations
  /locations/districts/{id}/microdistricts:
    get:
      consumes:
      - application/json
      description: Получает список микрорайонов для указанного района
      parameters:
      - description: ID района
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Microdistrict'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение микрорайонов района
      tags:
      - locations
  /locations/microdistricts:
    get:
      consumes:
      - application/json
      description: Получает список всех микрорайонов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Microdistrict'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение всех микрорайонов
      tags:
      - locations
  /locations/microdistricts/{id}:
    get:
      consumes:
      - application/json
      description: Получает информацию о микрорайоне по его идентификатору
      parameters:
      - description: ID микрорайона
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Microdistrict'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение микрорайона по ID
      tags:
      - locations
  /locations/regions:
    get:
      consumes:
      - application/json
      description: Получает список всех регионов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Region'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение всех регионов
      tags:
      - locations
  /locations/regions/{id}:
    get:
      consumes:
      - application/json
      description: Получает информацию о регионе по его идентификатору
      parameters:
      - description: ID региона
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Region'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение региона по ID
      tags:
      - locations
  /locations/regions/{id}/cities:
    get:
      consumes:
      - application/json
      description: Получает список городов для указанного региона
      parameters:
      - description: ID региона
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.City'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение городов региона
      tags:
      - locations
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"